Time: 37 ms (77.48%), Space: 11 MB (30.13%) - LeetHub

+++ 1209-remove-all-adjacent-duplicates-in-string-ii/1209-remove-all-adjacent-duplicates-in-string-ii.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
 string removeDuplicates(string s, int k) 
 {
-      vector<pair<char, short>> st;
+      vector<pair<char, int>> st;
       string res;
       for (auto ch : s)
       {
